Add files via upload
Scourgify.py is an application created to fix a data entry error in the before.csv file after it's already been completed, however the mistake went unnoticed. In the roster of students at Hogwarts, the first and last name were entered into the same row of data. This caused the file to only have 2 headers/rows instead of 3. The application reads in the file then corrects the error and writes a new CSV file with the mistake resolved. The three headers inserted are "first,last,house".